Refactor: Move Hub label test to test_hf_models.py and use config_args
